+++ test/pytest.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-pytest test/unit-test/test_cli_arg.py -v --cov=seedfarmer --cov-report=html --cov-report=xml
+pytest test/unit-test/*.py -v --cov=seedfarmer --cov-report=html --cov-report=xml